// src/components/Forms/Booking/tipo-demanda/interfaces.ts
export interface Autor {
  id: string;
  nome: string;
  documento: string;
  email: string;
  telefone: string;
}

export interface Reu {
  id: string;
  nome: string;
  documento: string;
}

export interface Pedido {
  id: string;
  descricao: string;
  valor: string;
}

export interface CamposAuditadosCadastroJuridico {
  data_audiencia: boolean | undefined;
  numero_reserva: boolean | undefined;
  lista_autores: boolean | undefined;
}

export interface CamposAuditadosProcon {
  data_defesa: boolean | undefined;
  data_audiencia: boolean | undefined;
  numero_reserva: boolean | undefined;
  lista_autores: boolean | undefined;
}
// ...
export interface DemandaProcon {
  pk?: string;
  usuario_atuando?: string;
  created_by?: string;
  cliente?: string;
  data_carimbo?: string;
  due_date?: string;
  arquivos?: ArquivoDemanda[];
  tipo_demanda?: string;
  status_demanda?: string;
  identificacao?: string;
  status?: string;

  campos_auditados?: CamposAuditadosProcon;

  numero_reserva?: string;
  data_defesa?: string;
  data_audiencia?: string;
  perfil_demanda?: string;
  tipificacao?: string;
  uf?: string;
  cidade?: string;
  objeto?: string;
  data_distribuicao?: string;
  data_reclamacao?: string;
  resumo_processo?: string;
  tipo_processo?: string;
  origem_reclamacao?: string;
  causa_raiz?: string;
  sub_causa_raiz?: string;
  lista_autores?: Autor[];
  lista_reus?: Reu[];
  observacao?: {
    mensagem: string;
    criada_em: string;
    criada_por: string;
  }[];
}